requirements.txt
MySQL-python==1.2.5
SQLAlchemy==1.1.9
pyzmq==2.1.7